Wire front-end to Http Server (#5433)
Update summary:
1. Send lobby login requests to http server
2. Send account creation requests to http server
 - Account creation flow is updated to be two steps
   1. Create account
   2. Re-use login flow with the newly created username+password
3. Wire moderator controls to (no-op) http server

Future updates will:
- implement moderator backend
- add moderator automated-host controls
- clean up lobby components

After this update a lobby instance is no longer needed/used, just http-server.

+++ http-clients/src/main/java/org/triplea/http/client/lobby/moderator/BanPlayerRequest.java
@@ -16,7 +16,7 @@ import org.triplea.domain.data.PlayerChatId;
 @Getter
 public class BanPlayerRequest {
   private String playerChatId;
-  private int banMinutes;
+  private long banMinutes;
 
   public PlayerChatId getPlayerChatId() {
     return PlayerChatId.of(playerChatId);
